+++ lib/server.dart
@@ -1,2 +1,3 @@
+export 'server/http_exve_api.dart';
 export 'server/http_server.dart';
-export 'server/http_handlers.dart';
+export 'server/http_simulator_api.dart';